moved drag and drop commands to separate file
